Add custom collision setting in the advanced skin builder

+++ common/src/main/java/moe/plushie/armourers_workshop/builder/client/gui/armourer/panel/ArmourerBaseSkinPanel.java
@@ -21,8 +21,8 @@ public class ArmourerBaseSkinPanel extends UIView {
 
     protected final SkinProperties skinProperties;
     protected final ArrayList<AbstractWidget> widgets = new ArrayList<>();
-    protected int cursorX = 0;
-    protected int cursorY = 0;
+    protected float cursorX = 0;
+    protected float cursorY = 0;
     protected Consumer<SkinProperties> applier;
 
     public ArmourerBaseSkinPanel(SkinProperties skinProperties) {
